requirements.txt
Django==5.1.7
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
django-auth-ldap==4.8.0
django-environ==0.11.2
django-solo==2.4.0
requests==2.32.3
htmx-django==1.0.0
whitenoise==6.8.2
gunicorn==23.0.0
